extract actions node into its own module
